Moving Send email to all users of a role action to RulesAction folder
